Сделал отображение звания в статистике, изменил функцию getUserTitle и вынёс из модуля в utils, улучшил функцию shuffleArray
