+++ src/components/user-title.js
@@ -1,10 +1,4 @@
-import {createElement} from '../utils.js';
-
-const userTitles = {
-  NOVICE: `Novice`,
-  FAN: `Fan`,
-  MOVIE_BUFF: `Movie Buff`,
-};
+import {createElement, getUserTitle} from '../utils.js';
 
 const createUserTitleTemplate = (films) => {
   const title = getUserTitle(films);
@@ -15,23 +9,6 @@ const createUserTitleTemplate = (films) => {
       <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
     </section>`
   );
-};
-
-const getUserTitle = (films) => {
-  const filmsInHistory = films.filter((it) => it.isInHistory).length;
-  let title;
-
-  if (filmsInHistory >= 1 && filmsInHistory <= 10) {
-    title = userTitles.NOVICE;
-  } else if (filmsInHistory >= 11 && filmsInHistory <= 20) {
-    title = userTitles.FAN;
-  } else if (filmsInHistory >= 21) {
-    title = userTitles.MOVIE_BUFF;
-  } else {
-    title = ``;
-  }
-
-  return title;
 };
 
 export default class UserTitle {
