+++ src/mock/film.js
@@ -216,12 +216,12 @@ const getRandomIntegerNumber = (min, max) => {
 };
 
 const shuffleArray = (array) => {
+
   for (let i = array.length - 1; i > 0; i--) {
     let j = Math.floor(Math.random() * (i + 1));
-    let temp = array[i];
-    array[i] = array[j];
-    array[j] = temp;
+    [array[i], array[j]] = [array[j], array[i]];
   }
+
   return array;
 };
